Virtual table improvements/fixes (#11)
* Fix error when there are no rows

* Don't clone the list of data provided

This isn't ideal, but it is a big performance win if the list is big. Since we don't make assumptions about the data outside of the rebuild method I think it's pretty safe.

* Update comments about perf

* Remove duplicated code that wires up handlers a second time

Possibly I merged/rebased badly - this code was originally moved, but we've ended up with two copies. This meant every event was being logged twice.