Rename row -> dataObject, rows -> data
To avoid confusion with table rows, the data is internally referred to as "data" (the whole list) and "dataObject" (a single object from that list of data).

The method for setting externally is still called setRows() since externally the table rows are not exposed and this feels logical, but we can always change it to setData if we'd like total consistency.