Port data model classes for the inspector to Dart. (#58)
* Port data model classes for the inspector to Dart.

This is a straight port copying the code and making it compile with Dart.
This is the a checkpoint. Next step is to actually run and test the code work
and strip out portions that are silly in Dart.

I'm splitting this CL into parts like this as this transform was mechanical
but the follow on CLs will contain new logic.