Tweaks to browser-launching service (#563)
* Allow ID to be any type

JSON-RPC allows for this, and other Flutter services handle strings.

* Change success flag for non-VM-service to simple bool

Success() ('type': 'success') seems to be a VM-Service convention but other stdio services like Flutter just have a bool.

* Ensure we handle all types of valid VM service URIs

Without this, an error occurs if you send a HTTP URI since it tries to connect to it as a websocket. This also adds /ws on the end if it's not there. This makes it valid to send the URL that's printed by the VM directly.

* Use the `method` field to decide how to handle the request

* Send both `event` and `method` for compatibility

We will remove the `method` one once VS Code has been updated for a while.

* Comment valid types for IDs

* Fix return type

* Ensure we always respond to vm.register even if URLs are invalid

* Fix comment

* Add a TODO

* Move to vm_service_lib-provided helper

Removes both copies of `getVmServiceUriFromObservatoryUri` and calls the shared `convertToWebSocketUrl` method.

* Fix error message