Replace usage of last private VM service protocol RPCs
The timeline page was using the private _getCpuProfileTimeline RPC which
is being removed from the VM and replaced with getCpuSamples. This new
RPC provides a minimally processed sample buffer which needs to be
further processed by clients.

This change updates 'getCpuProfileTimeline' in vm_service_wrapper.dart to
process the CpuSamples response from getCpuSamples into the format previously
returned by the _getCpuProfileTimeline RPC. In addition,
_clearCpuProfile was renamed to clearCpuSamples and made public.