[debugger] display contents of TypedData instances in debugger (#2153)
If a TypedData instance is displayed in the debugger, currently no contents are shown. Implement full support by decoding the provided bytes instance into the corresponding typed data. if bytes are provided but the kind instance is not recognized, then fallback to displaying raw bytes.

This does not currently display Uint64 or Int64 types due to limitations on the web.