[debugger] Add support for non-primitive map keys (#2154)
Currently a map with keys that do not have a meaningful valueAsString will just display [null]: in the dropdown. To support arbitrarily complex keys/maps, instead create an intermediate "Entry X" object, with a child bound variable for the key and value instanceRefs. This allows full inspection of key values.