Add support for DDS connections when using Devtools on web
As of 3.37 the VM service will attempt to redirect web socket connections to
DDS if DDS is connected. This works fine with some web socket implementations
(e.g., dart:io's WebSocket) but is unsupported in others (e.g., dart:html's
WebSocket). On the web we need to query the VM service via HTTP first to
determine which web socket URI we should use (see getWebSocketTarget
RPC for details).