Correctly process large HTTP responses in network profiler (#2602)
The network profiler was assuming that a response body would only ever
consist of a single chunk. This was resulting in us trying to decode a
partial response, causing a FormatException to be thrown and halting
processing of HTTP events.

This change stitches together all of the response body events for a
given response before trying to decode it. In addition, the
FormatException is caught and ignored in order to allow for non-UTF8
responses.

Fixes both https://github.com/flutter/devtools/issues/2398 and https://github.com/flutter/devtools/issues/2480