Throw RPCError when invoking getSourceReport in profile mode
There was a bug in the VM service where invoking getSourceReport on an
application running in profile mode (Dart AOT) would cause the VM to
crash.

This change adds a guard in vm_service_wrapper.dart to manually
check for a profile mode run before invoking the RPC, throwing the
correct RPCError that the VM service package would throw if the VM
didn't crash.