Handle nested scopes better when parsing code for syntax highlighting
Previously we would use the stack only to store the string scope names and would produce nested/overlapping tokens. With this change, we store more scope information on the stack and produce the tokens as scopes are pushed/popped so that the result is always a flat stream.

A new class was also added that makes it easier to pass around a "location" (an offset position, line number, and column number) because we now need to know the line/col for the ends of tokens (so that when we pop a token, we can then produce a span for the "lower layer" if it was nested).