Fix `CodeView`'s scrollable area
Currently, the `CodeView`'s scrollable area is only as wide as the
longest line in the currently selected script, resulting in arguably
unexpected behavior when trying to scroll while hovering over the right
side of the code area.

This change makes each line in the `CodeView` take up the maximum amount
of horizontal space possible, making the scrollable area consist of the
entire code viewport.