Add "Code Preview" section to Object Inspector views (#4447)
Adds support for displaying code associated with a given VM service
object. Dart code for only the given object will be displayed (e.g., a
subset of the containing script's source for functions, fields, and classes),
but will display the original range of line numbers.

In addition, users can use the "Code Preview" view to view and set
breakpoints on executable lines, just as if they were interacting with
the code view in the debugger.