Add "Code view" support to CPU Profiler
The current CPU profiler view renders CPU profiles in terms of Dart
functions. This is useful for the general Dart developer who is
interested in determining where time is being spent during execution,
but does not necessarily provide an accurate picture when profiling
performance of generated code. This is because compiler developers are
interested in the amount of time spent executing portions of generated
code, which can include inlined functions.

This PR adds an option in the CPU Profiler tab to display the CPU
profile without expanding inlined functions. When the "Code" view is
enabled, instead of building the CPU profile with a function tree the
tree is instead built based on `Code` objects from the VM.

This functionality is only enabled when VM developer mode is turned on,
and the extra data required to compute the code-based CPU profile is
only requested from the VM in this mode.

Related issue: https://github.com/flutter/devtools/issues/3861