Add initial support for non-query param navigation state (#4837)
This change adds support for providing additional state when navigating
through DevTools via `DevToolsRouterDelegate` without polluting the
query parameters with extra entries. This state is included in the
navigation history.

Controllers can mixin the new `RouteStateHandlerMixin` and override
`onRouteStateUpdate(DevToolsNavigationState)` to handle changes in the
current DevTools navigation state. To handle state changes,
`subscribeToRouterEvents` must first be called on the controller.

This change also adds support for jumping to a source location in the
debugger from URIs in the CPU profiler.