Patch and Dev Bump to start our new release version process (#4950)
Moving forward we will be doing daily dev version bumps. We have set up this process to fail if the master branch isn't on a -dev.\d+ version, so that the automation won't interfere if anyone is doing a release.

This bump puts the master branch version in a place where we can start doing those daily dev bumps.