Add the ability to enable platform types from the device list in the sidebar
For each unique platform type in the set of unsupported devices, this adds a new row ("Enable web for this project"). Clicking it triggers the same flow as VS Code (which currently involves a toast in the corner to confirm we can run `flutter create`).

In future we might want to change this prompt to be in the sidebar (so it's less likely you'll miss it).

Related PR for Dart-Code is at https://github.com/Dart-Code/Dart-Code/pull/4831

See https://github.com/Dart-Code/Dart-Code/issues/4820
See https://github.com/flutter/devtools/issues/6563