Don't render not-found page during initial setup
The router is always called with a `null` page when the app first loads. Previously this would render a not-found page and because there are also no args, it would be marked as not-embedded and would register with the server as a page that can be reused which (partly) caused https://github.com/Dart-Code/Dart-Code/issues/4832.