Add a global --flutter-from-path argument that overrides SDK selection
This promotes the `--flutter-from-path` flag to a global arg for all commands and uses it to locate the Flutter SDK before executing commands.

If set, we'll search for the SDK on PATH and fail if we don't find one.

If not set, we will infer an SDK from the one currently running the devtools_tool script (and again, fail if we're running from an SDK that doesn't have one).

When we call another instance of the tool, we bypass the shell script and instead execute it using Dart from the selected SDK (so that child processes use the selected SDK and we don't need to propagate the --flutter-from-path flag).

The update-flutter-sdk script was changed slightly to check whether the selected SDK is different to the tool/flutter-sdk one to decide whether to perform the extra update. This avoids some minor edge cases if:

- you have tool/flutter-sdk on PATH and pass the flag (it would update the Flutter SDK twice)
- you run devtools_tool from another SDK and don't pass the flag (it would not update the version being used to run the script)

Fixes https://github.com/flutter/devtools/issues/6867