Add extension type for old memory data JSON (#6978)
Work towards https://github.com/flutter/devtools/issues/6929

Very straightforward, least interesting new extension type. All this is is adding structure to the JSON types. It's hard to say that there is _great_ benefit to this change, but I do think it is a net positive. Happy to hear other thoughts.