Allow for null responses from `ServerConnectionStorage.getValue`
The `Storage` interface already specifies that the return type should be
nullable, but `ServerConnectionStorage` further restricted it to be
non-nullable. If a key was looked up and it did not have an entry in
storage, this would cause a type error and cause DevTools to hang on
start in some situations.

Fixes https://github.com/flutter/devtools/issues/7373