Add API fields to allow passing requiresDebugSession to openDevTools
This adds a new capability and parameter for openDevTools that allows DevTools to indicate to the editor that a call requires a debug session.

This allows calling openDevTools() without a debug session ID and letting the editor prompt so that things like Inspector could be top-level in the sidebar, but also supports having static tooling at the top level where this prompting is not desired.

If the `openDevToolsWithRequiresDebugSessionFlag` capability is `false`, this flag will not work and the behaviour (of whether an editor prompts or not) is not well defined, but for existing VS Code will prompt only for known DevTools pages that are not-static (that is, DevTools extensions will never prompt and always be assumed to be static if an explicit debug session is not passed).

This does not make any changes to the UI to use this functionality.