[third_party] Fix ansi_up compiling a regexp on every instance
Move the _csiRegexp to the top-level level to avoid re-compiling it
every time an `AnsiUp` instance is created and
`decodeAnsiColorEscapeCodes` is called.

Also move the comments out of the regexp string to avoid processing it
twice.

This library is not tested, but I've confirmed that the regexp string is
the same before and after the change. The regexp string:

    ^(?:\x1b\[([\x3c-\x3f]?)([\d;]*)([\x20-\x2f]?[\x40-\x7e]))|(?:\x1b\[[\x20-\x7e]*([\x00-\x1f:]))