Deprecate DTDConnectionInfo and replace with a class that can handle separate private/exposed URIs
This replaces `DTDConnectionInfo` with `DTDInfo` (unsure if a better name?) and marks the former as deprecated.

The new class has two URIs:

- `localUri` which is the URI used for backend services on the same machine as the DTD process (this is usually a `localhost` URI)
- `exposedUri` which is the URI that can be used from the frontend where the user is (this may or may not be the same as localUri depending on whether they're in a remote/web IDE session)

The DevTools web API returns the exposed URI (since it will be used by frontend tools) and other places that used the URI for connecting directly use the local one.

On its own, this change doesn't do anything, the intention is to roll this into the SDK and update the DevTools server code in dartdev/DDS to accept an additional optional URI which will then provide these back to the server code here (for the web API handler and extension search code).

This is work towards fixing some of the issues in https://github.com/Dart-Code/Dart-Code/issues/5158.