Delete unreachable `default` clause.
The Dart analyzer will soon be changed so that if the `default` clause
of a `switch` statement is determined to be unreachable by the
exhaustiveness checker, a new warning of type
`unreachable_switch_default` will be issued. This parallels the
behavior of the existing `unreachable_switch_case` warning, which is
issued whenever a `case` clause of a `switch` statement is determined
to be unreachable. For details see
https://github.com/dart-lang/sdk/issues/54575.

This PR deletes an unreachable `default` clause from `devtools` now,
to avoid a spurious warning when the analyzer change lands.