Remove the legacy postMessage version of the sidebar
(We should not merge this until after the release for the next stable)

This replaces the postMessage version of the sidebar (vsCodeFlutterPanel) with a message saying to upgrade to a newer version of the Dart extension. The last-but-one version of the Dart-Code extension uses the DTD version for current SDKs, so by the time this ships to stable you'll need to be ~6mo out of date with your Dart extension but on the latest SDK to see this.

This does not attempt any cleanup as a result of the removal, that will be done in another PR (or multiple) to keep it easy to review.