Normalize analytics paths for Windows
The analytics tests fail on Windows because it computes paths with backslashes but the test expects forward slashes.

While fixing, I figured it probably made sense to normalise both platforms paths to the same format so they could be grouped together more easily, so I made that change and updated the test to use the correct URI style depending on the current platform.

If this isn't right and we want to preserve the difference, we need to decide what the format of the stack traces should be when the URIs contain Windows paths.