Switch isDevToolsServerAvailable to not depend on the SSE connection
My recent change to disable the SSE connection when running embedded had an unintended side-effect. There's a number of places that check `isDevToolsServerAvailable` to see if they can call the HTTP API (and to set up server-side storage), however this was checking whether the SSE connection was made, not only that the server was available.

This change moves the server API check out of `DevToolsServerConnection` (which is basically the SSE API) and sets the storage (and therefore `isDevToolsServerAvailable`) based on just the HTTP API regardless of SSE, and then only attempts to connect SSE after.

Fixes https://github.com/flutter/devtools/issues/9023