Fix unsupported devices not showing "Enable x" if there are no other devices
In the list of devices we include devices that are not supported/enabled for the current platform, allowing you to enable them. However, the logic here would skip the entire device table if there are no enabled/supported devices, which would cause nothing to show up if there are none already enabled.

This changes the logic so the table is always shown if there are either kind of device, and only shows the help text if none are enabled.

Fixes https://github.com/Dart-Code/Dart-Code/issues/5661