update vm_service_protos package to version 2 (#9540)
Update to vm_service_protos version ^2.0.0, this allows us to update protobuf from 3.1.0 -> 4.2.0 (this is only a transitive dep).

Should unlock the fix for https://github.com/flutter/devtools/issues/9534, although we still need a new version of `vm_service_protos` to be published which allows `protobuf` ^5.0.0 (we need 5.1.0 to get the default size increase).