Fixes null error when parsing universal link settings (#9581)
fixes https://github.com/flutter/devtools/issues/8972

it is possible the tool can't get team id from the xcode project if the xcode developer team is not properly set up. In that case it will be null in the universal link setting's json. This Pr handles that cases

![build.yaml badge]

If you need help, consider asking for help on [Discord].

[build.yaml badge]: https://github.com/flutter/devtools/actions/workflows/build.yaml/badge.svg